+++ crates/net/network-api/Cargo.toml
@@ -18,6 +18,9 @@ alloy-rpc-types-admin.workspace = true
 reth-network-peers.workspace = true
 reth-network-types.workspace = true
 reth-network-p2p.workspace = true
+reth-eth-wire-types.workspace = true
+reth-tokio-util.workspace = true
+reth-ethereum-forks.workspace = true
 
 # ethereum
 alloy-primitives.workspace = true
@@ -27,6 +30,7 @@ enr = { workspace = true, default-features = false, features = ["rust-secp256k1"
 
 # async
 futures.workspace = true
+tokio-stream.workspace = true
 
 # misc
 thiserror.workspace = true
