+++ crates/net/network/benches/bench.rs
@@ -3,7 +3,7 @@ use criterion::*;
 use futures::StreamExt;
 use pprof::criterion::{Output, PProfProfiler};
 use rand::thread_rng;
-use reth_network::{test_utils::Testnet, NetworkEvents};
+use reth_network::{test_utils::Testnet, NetworkEventListenerProvider};
 use reth_network_api::Peers;
 use reth_primitives::U256;
 use reth_provider::test_utils::{ExtendedAccount, MockEthProvider};
