+++ crates/net/network/src/discovery.rs
@@ -15,6 +15,7 @@ use reth_discv5::{DiscoveredPeer, Discv5};
 use reth_dns_discovery::{
     DnsDiscoveryConfig, DnsDiscoveryHandle, DnsDiscoveryService, DnsNodeRecordUpdate, DnsResolver,
 };
+use reth_network_api::{DiscoveredEvent, DiscoveryEvent};
 use reth_network_peers::{NodeRecord, PeerId};
 use reth_network_types::PeerAddr;
 use reth_primitives::{EnrForkIdEntry, ForkId};
@@ -26,7 +27,6 @@ use tracing::trace;
 use crate::{
     cache::LruMap,
     error::{NetworkError, ServiceKind},
-    manager::DiscoveredEvent,
 };
 
 /// Default max capacity for cache of discovered peers.
@@ -326,15 +326,6 @@ impl Discovery {
             discovery_listeners: Default::default(),
         }
     }
-}
-
-/// Events produced by the [`Discovery`] manager.
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub enum DiscoveryEvent {
-    /// Discovered a node
-    NewNode(DiscoveredEvent),
-    /// Retrieved a [`ForkId`] from the peer via ENR request, See <https://eips.ethereum.org/EIPS/eip-868>
-    EnrForkId(PeerId, ForkId),
 }
 
 #[cfg(test)]
