+++ crates/net/network/src/lib.rs
@@ -139,7 +139,9 @@ mod swarm;
 
 pub use reth_eth_wire::{DisconnectReason, HelloMessageWithProtocols};
 pub use reth_network_api::{
-    BlockDownloaderProvider, NetworkInfo, Peers, PeersHandleProvider, PeersInfo,
+    BlockDownloaderProvider, DiscoveredEvent, DiscoveryEvent, NetworkEvent,
+    NetworkEventListenerProvider, NetworkInfo, PeerRequest, PeerRequestSender, Peers,
+    PeersHandleProvider, PeersInfo,
 };
 pub use reth_network_p2p::sync::{NetworkSyncUpdater, SyncState};
 pub use reth_network_types::{PeersConfig, SessionsConfig};
@@ -151,12 +153,11 @@ pub use session::{
 
 pub use builder::NetworkBuilder;
 pub use config::{NetworkConfig, NetworkConfigBuilder};
-pub use discovery::{Discovery, DiscoveryEvent};
+pub use discovery::Discovery;
 pub use fetch::FetchClient;
 pub use flattened_response::FlattenedResponse;
-pub use manager::{DiscoveredEvent, NetworkEvent, NetworkManager};
-pub use message::PeerRequest;
+pub use manager::NetworkManager;
 pub use metrics::TxTypesCounter;
-pub use network::{FullNetwork, NetworkEvents, NetworkHandle, NetworkProtocols};
+pub use network::{FullNetwork, NetworkHandle, NetworkProtocols};
 pub use swarm::NetworkConnectionState;
 pub use transactions::{FilterAnnouncement, MessageFilter, ValidateTx68};
