+++ crates/net/network/src/manager.rs
@@ -31,14 +31,13 @@ use futures::{Future, StreamExt};
 use parking_lot::Mutex;
 use reth_eth_wire::{
     capability::{Capabilities, CapabilityMessage},
-    DisconnectReason, EthVersion, Status,
+    DisconnectReason,
 };
 use reth_fs_util::{self as fs, FsPathError};
 use reth_metrics::common::mpsc::UnboundedMeteredSender;
-use reth_network_api::{EthProtocolInfo, NetworkStatus, PeerInfo};
+use reth_network_api::{EthProtocolInfo, NetworkEvent, NetworkStatus, PeerInfo, PeerRequest};
 use reth_network_peers::{NodeRecord, PeerId};
-use reth_network_types::{PeerAddr, PeersHandle, ReputationChangeKind};
-use reth_primitives::ForkId;
+use reth_network_types::{PeersHandle, ReputationChangeKind};
 use reth_storage_api::BlockNumReader;
 use reth_tasks::shutdown::GracefulShutdown;
 use reth_tokio_util::EventSender;
@@ -55,7 +54,7 @@ use crate::{
     eth_requests::IncomingEthRequest,
     import::{BlockImport, BlockImportOutcome, BlockValidation},
     listener::ConnectionListener,
-    message::{NewBlockMessage, PeerMessage, PeerRequest, PeerRequestSender},
+    message::{NewBlockMessage, PeerMessage},
     metrics::{DisconnectMetrics, NetworkMetrics, NETWORK_POOL_TRANSACTIONS_SCOPE},
     network::{NetworkHandle, NetworkHandleMessage},
     peers::PeersManager,
@@ -1043,67 +1042,6 @@ impl Future for NetworkManager {
 
         Poll::Pending
     }
-}
-
-/// (Non-exhaustive) Events emitted by the network that are of interest for subscribers.
-///
-/// This includes any event types that may be relevant to tasks, for metrics, keep track of peers
-/// etc.
-#[derive(Debug, Clone)]
-pub enum NetworkEvent {
-    /// Closed the peer session.
-    SessionClosed {
-        /// The identifier of the peer to which a session was closed.
-        peer_id: PeerId,
-        /// Why the disconnect was triggered
-        reason: Option<DisconnectReason>,
-    },
-    /// Established a new session with the given peer.
-    SessionEstablished {
-        /// The identifier of the peer to which a session was established.
-        peer_id: PeerId,
-        /// The remote addr of the peer to which a session was established.
-        remote_addr: SocketAddr,
-        /// The client version of the peer to which a session was established.
-        client_version: Arc<str>,
-        /// Capabilities the peer announced
-        capabilities: Arc<Capabilities>,
-        /// A request channel to the session task.
-        messages: PeerRequestSender,
-        /// The status of the peer to which a session was established.
-        status: Arc<Status>,
-        /// negotiated eth version of the session
-        version: EthVersion,
-    },
-    /// Event emitted when a new peer is added
-    PeerAdded(PeerId),
-    /// Event emitted when a new peer is removed
-    PeerRemoved(PeerId),
-}
-
-/// Represents events related to peer discovery in the network.
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub enum DiscoveredEvent {
-    /// Indicates that a new peer has been discovered and queued for potential connection.
-    ///
-    /// This event is generated when the system becomes aware of a new peer
-    /// but hasn't yet established a connection.
-    ///
-    /// # Fields
-    ///
-    /// * `peer_id` - The unique identifier of the discovered peer.
-    /// * `addr` - The network address of the discovered peer.
-    /// * `fork_id` - An optional identifier for the fork that this peer is associated with. `None`
-    ///   if the peer is not associated with a specific fork.
-    EventQueued {
-        /// The unique identifier of the discovered peer.
-        peer_id: PeerId,
-        /// The network address of the discovered peer.
-        addr: PeerAddr,
-        /// An optional identifier for the fork that this peer is associated with.
-        /// `None` if the peer is not associated with a specific fork.
-        fork_id: Option<ForkId>,
-    },
 }
 
 #[derive(Debug, Default)]
