+++ crates/net/network/src/state.rs
@@ -16,6 +16,7 @@ use rand::seq::SliceRandom;
 use reth_eth_wire::{
     capability::Capabilities, BlockHashNumber, DisconnectReason, NewBlockHashes, Status,
 };
+use reth_network_api::{DiscoveredEvent, DiscoveryEvent, PeerRequest, PeerRequestSender};
 use reth_network_peers::PeerId;
 use reth_network_types::{PeerAddr, PeerKind};
 use reth_primitives::{ForkId, B256};
@@ -24,13 +25,9 @@ use tracing::{debug, trace};
 
 use crate::{
     cache::LruCache,
-    discovery::{Discovery, DiscoveryEvent},
+    discovery::Discovery,
     fetch::{BlockResponseOutcome, FetchAction, StateFetcher},
-    manager::DiscoveredEvent,
-    message::{
-        BlockRequest, NewBlockMessage, PeerRequest, PeerRequestSender, PeerResponse,
-        PeerResponseResult,
-    },
+    message::{BlockRequest, NewBlockMessage, PeerResponse, PeerResponseResult},
     peers::{PeerAction, PeersManager},
     FetchClient,
 };
@@ -545,28 +542,30 @@ pub(crate) enum StateAction {
 
 #[cfg(test)]
 mod tests {
-    use crate::{
-        discovery::Discovery,
-        fetch::StateFetcher,
-        message::PeerRequestSender,
-        peers::PeersManager,
-        state::{BlockNumReader, NetworkState},
-        PeerRequest,
-    };
-    use reth_eth_wire::{
-        capability::{Capabilities, Capability},
-        BlockBodies, EthVersion,
-    };
-    use reth_network_p2p::{bodies::client::BodiesClient, error::RequestError};
-    use reth_network_peers::PeerId;
-    use reth_primitives::{BlockBody, Header, B256};
-    use reth_provider::test_utils::NoopProvider;
     use std::{
         future::poll_fn,
         sync::{atomic::AtomicU64, Arc},
     };
+
+    use reth_eth_wire::{
+        capability::{Capabilities, Capability},
+        BlockBodies, EthVersion,
+    };
+    use reth_network_api::PeerRequestSender;
+    use reth_network_p2p::{bodies::client::BodiesClient, error::RequestError};
+    use reth_network_peers::PeerId;
+    use reth_primitives::{BlockBody, Header, B256};
+    use reth_provider::test_utils::NoopProvider;
     use tokio::sync::mpsc;
     use tokio_stream::{wrappers::ReceiverStream, StreamExt};
+
+    use crate::{
+        discovery::Discovery,
+        fetch::StateFetcher,
+        peers::PeersManager,
+        state::{BlockNumReader, NetworkState},
+        PeerRequest,
+    };
 
     /// Returns a testing instance of the [`NetworkState`].
     fn state() -> NetworkState {
