+++ crates/net/network/src/test_utils/testnet.rs
@@ -12,7 +12,9 @@ use futures::{FutureExt, StreamExt};
 use pin_project::pin_project;
 use reth_chainspec::MAINNET;
 use reth_eth_wire::{protocol::Protocol, DisconnectReason, HelloMessageWithProtocols};
-use reth_network_api::{NetworkInfo, Peers, PeersHandleProvider};
+use reth_network_api::{
+    NetworkEvent, NetworkEventListenerProvider, NetworkInfo, Peers, PeersHandleProvider,
+};
 use reth_network_peers::PeerId;
 use reth_network_types::PeersHandle;
 use reth_provider::test_utils::NoopProvider;
@@ -39,8 +41,7 @@ use crate::{
     eth_requests::EthRequestHandler,
     protocol::IntoRlpxSubProtocol,
     transactions::{TransactionsHandle, TransactionsManager, TransactionsManagerConfig},
-    NetworkConfig, NetworkConfigBuilder, NetworkEvent, NetworkEvents, NetworkHandle,
-    NetworkManager,
+    NetworkConfig, NetworkConfigBuilder, NetworkHandle, NetworkManager,
 };
 
 /// A test network consisting of multiple peers.
